+++ packages/roosterjs-content-model-plugins/test/picker/PickerPluginTest.ts
@@ -1,0 +1,1 @@
+describe('PickerPlugin', () => {});
