+++ packages/roosterjs-content-model-plugins/test/picker/getQueryStringTest.ts
@@ -1,0 +1,3 @@
+describe('getQueryString', () => {
+    it('empty model');
+});
